backend/requirements.txt
django
psycopg2
djangorestframework
django-environ
django-cors-headers
djangorestframework-simplejwt
django-filter